Fix return type of QueryInterface::execute()
The `execute()` method has a boolean argument that
defines if a raw result should be returned.
This means that no object mapping is done but
an array containing the result values is returned.

It is now also supported to not just pass constant
boolean values but a variable containing a boolean
to the execute function. In this case it's not clear
if a raw result or objects should be returned
so we do now return a union type containing both.